ref #14: Add scope guard wrapper
Added a wrapper class where we can put different opencl types and
guarantee cleanup through destructor.
Added this header to the opencl_api rather than utility, as we usually
need it there rather than in the utility, as the opencl_api.h should
be the main entry point of the api.