fix #17: Fix crash when using GPU on ubuntu
Fixed a crash occuring when trying to use a GPU on Ubuntu.
The crash occured because the new createQueueWithProperties does not
work when using older OpenCL api's.

A simple if statement has been added to fix this issue,
assigning the correct queue function based on the version of the
platform.